main view is done
