+++ To Do App/Assets/Assets.xcassets/uncheck_v2.dataset/Contents.json
@@ -1,0 +1,13 @@
+{
+  "data" : [
+    {
+      "filename" : "Checklist (1)..tiff",
+      "idiom" : "universal",
+      "universal-type-identifier" : "public.tiff"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
