+++ To Do App/Views/AddItemView.swift
@@ -1,0 +1,20 @@
+//
+//  NewItemView.swift
+//  To Do App
+//
+//  Created by 𝕄𝕒𝕥𝕧𝕖𝕪 ℙ𝕠𝕕𝕘𝕠𝕣𝕟𝕚𝕪 on 25.06.2024.
+//
+
+import SwiftUI
+
+struct NewItemView: View {
+    
+    
+    var body: some View {
+        Form {
+            Text("New Item")
+        }
+    }
+    
+    
+}
